Pathfinding improved, although needs polish
The algorithm is able to pathfind with no problems, except when the
player is in certain areas. This happens because the position of the
player is at the "top-left", and this causes the destination to be in
unwalkable tiles, what causes the algorithm not to be able to find
player. This could be fixed by sending to the pathfinding algorithm the
position of the center of the player (convert_to_center_player_pos(x,y))