Pause finished. Read description.
There is a bug where the player or ground enemies will sometimes enter
in an infinite jump loop if the game is paused at a certain part of the
jump itself, so they will just keep ascending forever. Also, module gui
is provisionally set as unactive after starting the game due to the
starting screen and buttons getting blit while the game is running.